+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `Nanook::Wallet#remove_account`.
 - Added `Nanook::Wallet#republish_blocks`.
 - Added `Nanook::Wallet#search_pending`.
+- Added `Nanook::Wallet#work`.
 - Added `Nanook::Block#account`.
 - Added `Nanook::Block#amount`.
 - Added `Nanook::Block#balance`.